dj_brain_system/requirements_backend.txt
APScheduler==3.6.3
asgiref==3.7.2
cachetools==4.2.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.10
django-bootstrap5==23.4
django-debug-toolbar==4.3.0
django-filter==23.5
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
install==1.3.5
langdetect==1.0.9
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
packaging==24.0
pandas==2.2.1
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.4.0
setuptools==69.2.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tornado==6.4
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
